% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Aggregate_CRT}
\alias{Aggregate_CRT}
\title{Aggregate multiple cluster randomized trial (CRT) records with identical location}
\usage{
Aggregate_CRT(trial, auxiliaries = NULL)
}
\arguments{
\item{trial}{a dataframe containing locations (x,y) and optionally numerators, and corresponding denominators}

\item{auxiliaries}{a vector of the names of auxiliary variables to be summed across locations}
}
\value{
A dataframe with unique locations in which the numerators and denominators are summed over the locations
}
\description{
\code{Aggregate_CRT} aggregates data for binomial outcomes a trial dataframe contains multiple records with the same location, and outputs a dataframe with unique locations.
}
