% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Anonymise_TrialSite}
\alias{Anonymise_TrialSite}
\title{Anonymise locations in a trial site}
\usage{
Anonymise_TrialSite(trial = NULL)
}
\arguments{
\item{trial}{dataframe with Cartesian co-ordinates of households (columns x and y)}
}
\value{
data.frame with modified co-ordinates of households (other columns remain unchanged)
}
\description{
\code{Anonymise_TrialSite} carries out rotation of x,y coordinates a random angle about a random origin. Coordinates centred on the origin are returned.
}
\examples{
#Rotate and reflect test site locations
transformedTestlocations <- Anonymise_TrialSite(trial = readdata('test_site.csv'))
}
