% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anonymize.site}
\alias{anonymize.site}
\title{Anonymize locations of a trial site}
\usage{
anonymize.site(trial, latvar = "lat", longvar = "long")
}
\arguments{
\item{trial}{\code{"CRTsp"} object or trial data frame with co-ordinates of
households}

\item{latvar}{name of column containing latitudes in decimal degrees}

\item{longvar}{name of column containing longitudes in decimal degrees}
}
\value{
A list of class \code{"CRTsp"} containing the following components:
}
\description{
\code{anonymize.site} Transforms coordinates to maintain confidentiality
}
\details{
The coordinates are transformed in order to obscure geographical information, to support confidentiality of
information linked to households. The input may have either \code{lat long} or \code{x,y} coordinates.
The function first searches for any \code{lat long} co-ordinates and converts these to \code{x,y}
Cartesian coordinates. These are then are rotated by a random angle about a random origin. The returned object
has transformed co-ordinates re-centred at the origin. Other data are unchanged.
}
\examples{
#Rotate and reflect test site locations
transformedTestlocations <- anonymize.site(trial =  readdata("exampleCRT.csv"))
}
