% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AvecNet_coordinates}
\alias{AvecNet_coordinates}
\title{Coordinates from the Avecnet trial}
\format{
dataframe:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
}
}
\usage{
AvecNet_coordinates
}
\description{
Coordinates from the Avecnet trial
}
\keyword{datasets}
