% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designCRT.R
\name{calculateCRTpower}
\alias{calculateCRTpower}
\title{Power and sample size calculations for a CRT}
\usage{
calculateCRTpower(
  locations,
  ICC,
  effect,
  alpha,
  outcome.type,
  sigma2 = NULL,
  phi = 1,
  N = 1,
  desiredPower,
  yC,
  mean_h,
  sd_h
)
}
\arguments{
\item{locations}{numeric: total number of units}

\item{ICC}{numeric: Intracluster Correlation Coefficient}

\item{effect}{numeric: Required effect size}

\item{alpha}{numeric: Significance level}

\item{sigma2}{numeric variance of the outcome (for \code{outcome.type = "y"})}

\item{phi}{overdispersion parameter (for \code{outcome.type = "n"} or \code{outcome.type = "e"})}

\item{N}{denominator for proportions (for \code{outcome.type = "p"})}

\item{desiredPower}{desired power (expressed as a proportion)}

\item{yC}{expected average outcome in control arm}

\item{mean_h}{mean number of units per cluster}

\item{sd_h}{standard deviation of number of units per cluster}

\item{outcome.type:}{\tabular{ll}{
\code{"y"}\tab continuous \cr
\code{"n"}\tab count \cr
\code{"e"}\tab event rate \cr
\code{"p"}\tab proportion  \cr
\code{"d"}\tab dichotomous \cr
}}
}
\value{
A CRT object comprising the input coordinates :
\itemize{
\item \code{input.parameters} list of input parameters
\item \code{CRT.design.full} list of calculated filename for coordinates of households
}
}
\description{
\code{calculateCRTpower} carries out power and sample-size calculations for a CRT
The output is a CRT object containing values for:
\itemize{
\item The required numbers of clusters to achieve the specified power.
\item The design effect based on the input ICC.
\item The power achievable.
}
}
\details{
for count data, or event rates a quasi–Poisson model is assumed
}
\examples{
examplePower = calculateCRTpower(locations = 3000, ICC=0.10, effect=0.4, alpha = 0.05,
    outcome.type = 'd', desiredPower = 0.8, yC=0.35, mean_h=100, sd_h=5)
}
