% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTanalysis_methods.R
\name{coef.CRTanalysis}
\alias{coef.CRTanalysis}
\title{Extract model coefficients}
\usage{
\method{coef}{CRTanalysis}(object, ...)
}
\arguments{
\item{object}{CRTanalysis object}

\item{...}{other arguments}
}
\value{
the model coefficients returned by the statistical model run within the \code{CRTanalysis} function
}
\description{
\code{coef.CRTanalysis} method for extracting model fitted values
}
\examples{
{example <- readdata('exampleCRT.txt')
exampleGEE <- CRTanalysis(example, method = "GEE")
coef(exampleGEE)
}
}
