% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Convert_LatLong}
\alias{Convert_LatLong}
\title{Convert lat long co-ordinates to x,y}
\usage{
Convert_LatLong(df, latvar = "lat", longvar = "long")
}
\arguments{
\item{df}{data frame containing latitudes and longitudes in decimal degrees}

\item{latvar}{name of column containing latitudes in decimal degrees}

\item{longvar}{name of column containing longitudes in decimal degrees}
}
\value{
The input dataframe with the lat-long coordinates replaced with Cartesian coordinates in units of km, centred on (0,0)
\itemize{
\item \code{x}: x co-ordinates
\item \code{y}: y co-ordinates
}
}
\description{
\code{Convert_LatLong} converts co-ordinates expressed as decimal degrees into x,y using the equirectangular projection (valid for small areas).
Coordinates centred on the origin are returned.
}
