% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CRT_as_data.frame}
\alias{CRT_as_data.frame}
\title{Convert object of S3 class CRT to a data frame}
\usage{
CRT_as_data.frame(CRT)
}
\arguments{
\item{CRT}{name of CRT object}
}
\value{
data frame with one row for each location
}
\description{
\code{CRT_as_data.frame} removes lists of descriptors and summary statistics
}
\details{
\code{CRT_as_data.frame} removes lists of descriptors and summary statistics
}
