% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTspat-package.R
\docType{package}
\name{CRTspat-package}
\alias{CRTspat}
\alias{CRTspat-package}
\title{CRTspat: Design and Analysis of Cluster Randomized Trials of Health Interventions with Spatial Contamination}
\description{
Supports design, workflow and analysis of Cluster Randomised Trials of (health) interventions where there may be geographical contamination between the arms.
}
\author{
\strong{Maintainer}: Thomas Smith \email{Thomas-a.Smith@unibas.ch} (\href{https://orcid.org/0000-0002-3650-9381}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Lea Multerer [contributor]
  \item Mariah Silkey [contributor]
}

}
\keyword{internal}
