% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTspillover-package.R
\docType{package}
\name{CRTspillover-package}
\alias{CRTspillover}
\alias{CRTspillover-package}
\title{CRTspillover: Design of CRTs for malaria control interventions with contamination}
\description{
CRTspillover comprises a suite of functions to support design of cluster randomised trials of (health) interventions where there may be geographical contamination between the arms. cluster size and sample size for cluster randomized controls of interventions with geographical contamination, and for carrying out randomizations for such trials. The methods are described in Multerer et al. (in preparation). The package takes as inputs a table of geolocations of experimental units. Example datasets are provided taken from field trials of malaria interventions.
}
\keyword{internal}
