% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTspillover.r
\name{CRTspillover}
\alias{CRTspillover}
\title{Designs, simulates, and analyses a cluster randomised trial (CRT) with geographical contamination}
\usage{
CRTspillover(
  alpha = 0.05,
  power = 0.8,
  effect = 0.4,
  ICC = 0.175,
  pC = 0.4,
  cont = 0.25,
  coordinates = AvecNet_coordinates,
  h = 80,
  algo = "kmeans",
  reuseTSP = FALSE,
  sd = 0.4,
  method = "GEE",
  excludeBuffer = FALSE,
  requireBootstrap = FALSE,
  resamples = 1000,
  iter = 10000,
  burnin = 5000
)
}
\arguments{
\item{alpha}{confidence level}

\item{power}{power}

\item{effect}{required effect size}

\item{ICC}{Intra-Cluster Correlation obtained from other studies}

\item{pC}{baseline prevalence}

\item{cont}{contamination range in km, obtained from other studies}

\item{coordinates}{dataframe containing coordinates of households. Columns 'x' and 'y' should contain Cartesian (x,y) coordinates. Units are expected to be km.}

\item{h}{proposal for the number of coordinates in each cluster}

\item{algo}{algorithm for cluster boundaries, choose between
"TSP": travelling salesman problem heuristic;
"NN": nearest neighbor;
"kmeans": kmeans algorithm}

\item{reuseTSP}{indicator of whether a pre-existing path should be used by the TSP algorithm}

\item{sd}{standard deviation measuring spatial smoothing of proportion}

\item{method}{statistical method used to analyse trial.
Options are 'piecewise_linear','logit','sigmoid','empirical','GEE','MCMC01','MCMC02','MCMC03'}

\item{excludeBuffer}{exclude any buffer zone (records with buffer=TRUE) from the analysis}

\item{requireBootstrap}{logical indicator of whether bootstrap confidence intervals are required}

\item{resamples}{number of bootstrap samples if bootstrap confidence intervals are required}

\item{iter}{total number of interations of MCMC algorithm (for MCMC methods)}

\item{burnin}{number of burnin interations of MCMC algorithm (for MCMC methods)}
}
\value{
list object with the following attributes:
\itemize{
\item \code{design}: list of vector-specific entomological parameters (see
description of output from function \code{Design_CRT})
\item \code{simulation}: vector of intervention-specific entomological
parameters and impact (vectorial capacity) values
\item \code{analysis}: list of results of statistical analysis of the simulated trial (see
description of output from function \code{Analyse_CRT})
}
}
\description{
\code{CRTspillover} estimates the required number of clusters and the extent of contamination between arms for a CRT based on an input set of locations.
The trial is simulated and the simulated data are analysed.
}
\examples{
generate a simulated trial and analyse by GEE
exampleAnalysis=CRTspillover()
}
