% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design_CRT.R
\name{Design_CRT}
\alias{Design_CRT}
\title{Design a CRT of a malaria intervention with contamination}
\usage{
Design_CRT(
  trial,
  alpha = 0.05,
  desiredPower = 0.8,
  effect,
  ICC,
  pC,
  buffer.width = 0,
  h,
  algo = "kmeans",
  outcome.type = outcome.type
)
}
\arguments{
\item{trial}{dataframe or CRT object containing Cartesian coordinates of locations in columns 'x' and 'y'.
Units are expected to be km.}

\item{alpha}{confidence level}

\item{desiredPower}{desired power}

\item{effect}{required effect size}

\item{ICC}{Intra-Cluster Correlation}

\item{pC}{baseline prevalence}

\item{buffer.width}{required buffer width in km}

\item{h}{proposal for the number of coordinates in each cluster}

\item{algo}{algorithm used to determine cluster boundaries, options are-
'TSP': travelling salesman problem heuristic;
'NN': nearest neighbor;
'kmeans': kmeans}

\item{outcome.type}{options are 'Continuous', 'Count or rate', 'Proportion', 'Dichotomous'}
}
\value{
A CRT object comprising the input data, cluster and arm assignments, trial description and results of power calculations
\itemize{
\item \code{arm}: arm assignment
\item \code{cluster}: cluster assignment
\item \code{buffer}: indicator of whether the location is in the buffer between arms
\item \code{input.parameters} list of input parameters
\item \code{CRT.design.full} list of calculated filename for coordinates of households
}
}
\description{
\code{Design_CRT} estimates the required number of clusters and the extent of contamination between arms for a CRT
with the input set of locations. Outputs are:
(i) Estimates of the required numbers of clusters.
(ii) A proposal for the cluster and arm assignments to the input coordinates.
(A warning is output if the number of locations is too small to allow randomisation of sufficient clusters).
(iii) the proportion of households in the input geography falling within the core of the clusters
(i.e. outside the contamination range of locations in the opposite arm)
}
\examples{

exampleDesign = Design_CRT(trial=readdata('test_site.csv'),
                ICC=0.10, effect=0.4, pC=0.35, desiredPower = 0.8, algo = "kmeans",
                buffer.width=0.25, h=100, outcome.type ="Dichotomous")
}
