% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_example_baseline}
\alias{get_example_baseline}
\title{Helper function to read in example dataset from csv format}
\usage{
get_example_baseline()
}
\value{
A data.frame comprising with the following columns:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
\item \code{RDT_test_result}: test result (0= negative, 1 = positive)
}
}
\description{
\code{get_example_baseline} reads the example dataset from .csv format
}
