% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{latlong_as_xy}
\alias{latlong_as_xy}
\title{Convert lat long co-ordinates to x,y}
\usage{
latlong_as_xy(df, latvar = "lat", longvar = "long")
}
\arguments{
\item{df}{data frame containing latitudes and longitudes in decimal degrees}

\item{latvar}{name of column containing latitudes in decimal degrees}

\item{longvar}{name of column containing longitudes in decimal degrees}
}
\value{
A list of S3 class \code{"CRT"} containing the following components:
\tabular{llll}{
\code{geom.full}   \tab list: \tab summary statistics describing the site \tab\cr
\code{trial} \tab data frame: \tab rows correspond to geolocated points, as follows:\tab\cr
\tab \code{x} \tab numeric vector: \tab x-coordinates of locations \cr
\tab \code{y} \tab numeric vector: \tab y-coordinates of locations \cr
\tab \code{...} \tab \tab other objects included in the input \code{"CRT"} object or data frame \cr
}
}
\description{
\code{latlong_as_xy} converts co-ordinates expressed as decimal degrees
into x,y
}
\details{
An object containing the input locations replaced with Cartesian
coordinates in units of km, centred on (0,0). Other data are unchanged.
The equirectangular projection (valid for small areas) is used.
}
