% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_CRTmap.R
\name{Plot_Contamination}
\alias{Plot_Contamination}
\title{\code{Plot_Contamination} returns a plot of an estimated contamination function along with
the grouped proportions of positives by distance from the arm boundary}
\usage{
Plot_Contamination(analysis)
}
\arguments{
\item{analysis}{analysis output by Analyse_CRT}
}
\value{
graphics object produced by the ggplot2 package
}
\description{
\code{Plot_Contamination} returns a plot of an estimated contamination function along with
the grouped proportions of positives by distance from the arm boundary
}
\examples{
#Plot locations only
#Plot clusters in colour
}
