% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_CRTmap.R
\name{Plot_CRTmap}
\alias{Plot_CRTmap}
\title{Map of CRT trial area}
\usage{
Plot_CRTmap(
  trial = trial,
  showLocations = TRUE,
  showClusterBoundaries = TRUE,
  showClusterLabels = TRUE,
  colourClusters = TRUE,
  showArms = TRUE,
  analysis = NULL,
  showDistance = FALSE,
  showProportion = FALSE,
  cpalette = NULL,
  maskbuffer = 0.2,
  labelsize = 4
)
}
\arguments{
\item{trial}{standard deviation of random displacement from each settlement cluster center}

\item{showLocations}{logical: determining whether household locations are shown}

\item{showClusterBoundaries}{logical: determining whether clusters are shown}

\item{showClusterLabels}{logical: determining whether clusters are numbered}

\item{colourClusters}{logical: determining whether clusters are coloured}

\item{showArms}{logical: determining whether the areas assigned to each trial arm are shown}

\item{analysis}{list: analysis object generated by Analyse_CRT using method equal to one of 'LR','CRE','SPDE','SPCRE'}

\item{showDistance}{logical: determining whether smoothed raster of the distance from nearest discordant location is produced}

\item{showProportion}{logical: determining whether a smoothed raster of the fitted proportion is produced}

\item{cpalette}{colour palette (to use different colours for clusters must be at least as long as the number of clusters, defaults to rainbow())}

\item{maskbuffer}{radius of buffer drawn around inhabited areas}

\item{labelsize}{size of labels giving cluster numbers}
}
\value{
graphics object produced by the ggplot2 package
}
\description{
\code{Plot_CRTmap} returns a graphics object created using ggplot2.
Cartesian (x,y) coordinates are used. Units are expected to be km.
}
\examples{
#Plot locations only
Plot_CRTmap(trial=testArms,showArms=FALSE,showClusterBoundaries=FALSE,
           colourClusters=FALSE, maskbuffer=0.2)

#Plot clusters in colour
Plot_CRTmap(trial=testArms, showArms=FALSE, colourClusters=TRUE, labelsize=2, maskbuffer=0.2)

#Plot arms
Plot_CRTmap(trial=testArms, maskbuffer=0.2)

#Plot distance to nearest discordant location

}
