% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_CRTmap.R
\name{Plot_DataByDistance}
\alias{Plot_DataByDistance}
\title{\code{Plot_DataByDistance} returns a stacked bar chart of the grouped data
the grouped proportions of positives by distance from the arm boundary}
\usage{
Plot_DataByDistance(
  trial = trial,
  num = num,
  denom = denom,
  cpalette = c("#D55E00", "#009E73", "#0072A7", "#C879C8")
)
}
\arguments{
\item{trial}{a dataframe containing locations (x,y), cluster assignments, and arm assignments}

\item{num}{name of numerator variable}

\item{denom}{name of denominator variable}

\item{cpalette}{colour palette (to use different colours for clusters must be at least as long as the number of clusters, defaults to rainbow())}
}
\value{
graphics object produced by the ggplot2 package
}
\description{
\code{Plot_DataByDistance} returns a stacked bar chart of the grouped data
the grouped proportions of positives by distance from the arm boundary
}
\examples{
#Plot locations only
#Plot clusters in colour
}
