% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCRT.R
\name{plotContamination}
\alias{plotContamination}
\title{\code{plotContamination} returns a plot of an estimated contamination function}
\usage{
plotContamination(analysis)
}
\arguments{
\item{analysis}{list object of class CRT produced by \code{analyseCRT()}}
}
\value{
graphics object produced by the ggplot2 package
}
\description{
\code{plotContamination} returns a plot of an estimated contamination function
}
\details{
The fitted contamination function is plotted as a continous blue line against the distance from the nearest discordant
location.Using the same axes, data summaries are plotted for ten categories of distance from the boundary. Both the
average of the outcome and confidence intervals are plotted. \cr
For analyses with logit link function the outcome is plotted as a proportion. \cr
For analyses with log link function the outcome is plotted on a scale of the Williams' mean
(mean of exp(log(x + 1))) - 1) \cr
}
\examples{
plotContamination(analysis = readdata('test_Analyse_CRT.txt'))
}
