% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCRT.R
\name{plotCRTmap}
\alias{plotCRTmap}
\title{Map of CRT trial area}
\usage{
plotCRTmap(
  trial = trial,
  fill = "arms",
  showLocations = FALSE,
  showClusterBoundaries = TRUE,
  showClusterLabels = FALSE,
  showContamination = FALSE,
  cpalette = NULL,
  maskbuffer = 0.2,
  labelsize = 4,
  legend.position = NULL
)
}
\arguments{
\item{trial}{either:\cr
an object of class \code{"CRT"}; \cr
a data frame containing locations in (x,y) coordinates,
(optionally) cluster assignments (factor \code{cluster}), and arm assignments (factor \code{arm});\cr
an object of class \code{CRTanalysis} produced by an \code{"INLA"} analysis.}

\item{fill}{fill layer of map
\tabular{ll}{
\code{"cluster"} \tab cluster assignment \cr
\code{"arms"}   \tab arm assignment \cr
\code{"distance"} \tab distance to the nearest discordant location\cr
\code{"prediction"}\tab model prediction of the outcome \cr
\code{"none"}\tab No fill \cr
}}

\item{showLocations}{logical: determining whether locations are shown}

\item{showClusterBoundaries}{logical: determining whether cluster boundaries are shown}

\item{showClusterLabels}{logical: determining whether the cluster numbers are shown}

\item{showContamination}{logical: determining whether the estimated contamination range should be shown}

\item{cpalette}{colour palette (to use different colours for clusters must be at least as long as the number of clusters, defaults to rainbow())}

\item{maskbuffer}{radius of buffer drawn around inhabited areas (km)}

\item{labelsize}{size of cluster number labels}

\item{legend.position}{(using ggplot2::themes syntax)}
}
\value{
graphics object produced by the ggplot2 package
}
\description{
\code{plotCRTmap} returns a graphics object created using ggplot2.
Cartesian (x,y) coordinates are used. Units are expected to be km.
}
\examples{
#Plot locations only
plotCRTmap(trial = readdata('test_CRT2.csv'), fill = "none", showLocations = TRUE,
           showClusterBoundaries=FALSE, maskbuffer=0.2)

#Show cluster boundaries and number clusters
plotCRTmap(trial = readdata('test_CRT2.csv'), fill ="none", showClusterBoundaries=TRUE,
           showClusterLabels=TRUE, maskbuffer=0.2)

#Plot clusters in colour
plotCRTmap(trial=readdata('test_CRT2.csv'), fill = "clusters", showClusterLabels = TRUE,
          labelsize=2, maskbuffer=0.2)

#Plot arms
plotCRTmap(trial=readdata('test_CRT2.csv'), maskbuffer=0.2, legend.position=c(0.2,0.8))

}
