% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCRT.R
\name{plotDataByDistance}
\alias{plotDataByDistance}
\title{\code{plotDataByDistance} returns a stacked bar chart of the grouped data
the outcome grouped by distance from the arm boundary}
\usage{
plotDataByDistance(
  trial = trial,
  num = num,
  denom = denom,
  cpalette = c("#D55E00", "#0072A7")
)
}
\arguments{
\item{trial}{a dataframe containing locations (x,y), cluster assignments, and arm assignments}

\item{num}{name of numerator variable}

\item{denom}{name of denominator variable}

\item{cpalette}{colour palette (to use different colours for clusters must be at least as long as the number of clusters, defaults to rainbow())}
}
\value{
graphics object produced by the ggplot2 package
}
\description{
\code{plotDataByDistance} returns a stacked bar chart of the grouped data
the outcome grouped by distance from the arm boundary
}
\examples{
plotDataByDistance(trial=readdata('test_Simulate_CRT.csv'))
}
