% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Randomize_CRT}
\alias{Randomize_CRT}
\title{Randomize a two-armed cluster randomized trial}
\usage{
Randomize_CRT(
  trial,
  matchedPair = FALSE,
  baselineNumerator = "base_num",
  baselineDenominator = "base_denom"
)
}
\arguments{
\item{trial}{a \code{CRT} object or \code{data.frame} containing locations in (x,y) coordinates; cluster
assignments (factor \code{cluster}); optionally: specification of a buffer zone (logical \code{buffer});
any other variables required for subsequent analysis.}

\item{matchedPair}{logical: indicator of whether pair-matching on the
baseline data should be used in randomization}

\item{baselineNumerator}{name of numerator variable for baseline data (required for
matched-pair randomization)}

\item{baselineDenominator}{name of denominator variable for baseline data (required for
matched-pair randomization)}
}
\value{
A list of class \code{CRT} containing the following components:
\tabular{llll}{
\code{CRT.design.full}:   \tab list \tab summary statistics of the site \cr
\code{CRT.design.core}:   \tab list \tab summary statistics of the core area (if a buffer is present)\cr
\code{x}: \tab numeric vector \tab x-coordinates of locations \cr
\code{y}: \tab numeric vector \tab y-coordinates of locations \cr
\code{cluster}: \tab factor \tab assignments to cluster of each location  \cr
\code{pair}: \tab factor \tab assignments to matched pair of each location  \cr
(if \code{matchedPair} randomisation was carried out) \cr
\code{arm}: \tab factor \tab assignments to \code{control} or \code{intervention} for each location \cr
\code{...};   \tab other objects included in the input \code{CRT} object or \code{data.frame}  \cr
}
}
\description{
\code{Randomize_CRT} carries out randomization of clusters for a CRT and
augments the trial dataframe with assignments to arms \cr
}
\examples{
#Randomize the clusters in an example trial
set.seed(352)
example.CRT <- Randomize_CRT(trial = readdata('test_Clusters.csv'), matchedPair = TRUE)
}
