% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readdata}
\alias{readdata}
\title{Read test dataset}
\usage{
readdata(filename)
}
\arguments{
\item{filename}{name of text file stored within the package}
}
\value{
R object corresponding to the text file
}
\description{
\code{readdata} reads a file from package library of .csv or .txt test datasets
}
