% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseCRT.R
\name{residuals.CRTanalysis}
\alias{residuals.CRTanalysis}
\title{Extract model residuals}
\usage{
\method{residuals}{CRTanalysis}(object, ...)
}
\arguments{
\item{object}{CRTanalysis object}

\item{...}{other arguments}
}
\value{
the residuals from the statistical model run within the \code{CRTanalysis} function
}
\description{
\code{residuals.CRTanalysis} method for extracting model residuals
}
\examples{
{example <- readdata('exampleCRT.txt')
exampleGEE <- CRTanalysis(example, method = "GEE")
residuals <- residuals(exampleGEE)
}
}
