% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_TrialSite.R
\name{Simulate_TrialSite}
\alias{Simulate_TrialSite}
\title{Create a set of Cartesian co-ordinates for a simulated CRT}
\usage{
Simulate_TrialSite(scale = 0.25, households = 2500, kappa = 4, mu = 50)
}
\arguments{
\item{scale}{standard deviation of random displacement from each settlement cluster center}

\item{households}{number of households in population}

\item{kappa}{intensity of Poisson process of settlement cluster centers}

\item{mu}{mean  number of points per settlement cluster}
}
\value{
data frame with co-ordinates of households
}
\description{
\code{Simulate_TrialSite} Create a set of Cartesian co-ordinates using the Thomas algorithm to simulate a human settlement pattern
}
\examples{
#Generate a simulated area with 10,000 households
example_area = Simulate_TrialSite(scale = 2, households=10000, kappa=3, mu=40)
}
