% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCRT.R
\name{simulateSite}
\alias{simulateSite}
\title{Create co-ordinates for a simulated CRT}
\usage{
simulateSite(geoscale, locations, kappa, mu)
}
\arguments{
\item{geoscale}{standard deviation of random displacement from each settlement cluster center}

\item{locations}{number of locations in population}

\item{kappa}{intensity of Poisson process of settlement cluster centers}

\item{mu}{mean  number of points per settlement cluster}
}
\value{
A list with class \code{CRT} containing the following components:
\tabular{llll}{
\code{CRT.design.full}   \tab list \tab summary statistics describing the site\cr
\code{x} \tab numeric vector \tab x-coordinates of locations \cr
\code{y} \tab numeric vector \tab y-coordinates of locations \cr
}
}
\description{
\code{simulateSite} creates a set of Cartesian co-ordinates for use as the locations in a simulated trial site
}
\details{
\code{simulateSite} simulates a human settlement pattern using the Thomas algorithm (\code{rThomas} function
in \href{http://spatstat.org/}{\code{spatstat}} allowing the user to defined the density of locations and degree of spatial clustering.
The results are output as a set of Cartesian coordinates centred at the origin.
}
\examples{
# Generate a simulated area with 10,000 locations
example_area = simulateSite(geoscale = 1, locations=10000, kappa=3, mu=40)
}
