% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Solarmal_baseline}
\alias{Solarmal_baseline}
\title{Baseline RDT positivity data from the SolarMal trial}
\format{
data.frame:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
\item \code{RDT_test_result}: test result (0= negative, 1 = positive)
\item \code{X}: row number
}
}
\usage{
Solarmal_baseline
}
\description{
Baseline RDT positivity data from the SolarMal trial
}
\keyword{datasets}
