% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{specify.clusters}
\alias{specify.clusters}
\title{algorithmically assigns locations to clusters in a CRT}
\usage{
specify.clusters(
  trial = trial,
  h = NULL,
  nclusters = NULL,
  algorithm = "NN",
  reuseTSP = FALSE
)
}
\arguments{
\item{trial}{A CRT object or data frame containing (x,y) coordinates of
households}

\item{h}{integer: proposal for the number of locations in each cluster}

\item{nclusters}{integer: number of clusters}

\item{algorithm}{algorithm for cluster boundaries, options are:
\code{NN} (the default),  \code{kmeans},  \code{TSP}}

\item{reuseTSP}{logical: indicator of whether a pre-existing path should be used by
the TSP algorithm}
}
\value{
A list of class \code{"CRT"} containing the following components:
\tabular{lll}{
\code{CRT.design.full}   \tab list \tab summary statistics describing the site and cluster assignments.\cr
\code{x} \tab numeric vector \tab x-coordinates of locations \cr
\code{y} \tab numeric vector \tab y-coordinates of locations \cr
\code{cluster} \tab factor \tab assignments to cluster of each location  \cr
\code{path} \tab numeric \tab travelling salesman path (created only if the TSP algorithm is used) \cr
\code{...};   \tab other objects included in the input \code{"CRT"} object or data frame  \cr
}
}
\description{
\code{specify.clusters} algorithmically assigns locations to clusters by grouping them geographically
}
\details{
Clustering is carried out using one of three algorithms:
\tabular{lll}{
\code{NN}\tab Nearest neighbour:\tab assigns equal numbers of locations to each cluster \cr
\code{kmeans}\tab kmeans clustering  :\tab aims to partition locations so that each
belongs to the cluster with the nearest centroid \cr
\code{TSP} \tab travelling salesman problem heuristic :\tab Assigns locations sequentially
along a travelling salesman path.The \code{reuseTSP} parameter is used to allow the path to be reused
for creating alternative allocations with different cluster sizes. \cr
}
Either \code{nclusters} or \code{h} must be specified. If both are specified
the input value of \code{nclusters} is ignored.\cr
}
\examples{
#Assign clusters to the test trial dataset averaging h = 40 using the kmeans algorithm
exampletrial <- specify.clusters(trial = readdata('test_site.csv'),
                            h = 40, algorithm = 'kmeans', reuseTSP = FALSE)
}
