% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{specify_clusters}
\alias{specify_clusters}
\title{Algorithmically assign locations to clusters in a CRT}
\usage{
specify_clusters(
  trial = trial,
  k = NULL,
  h = NULL,
  algorithm = "NN",
  reuseTSP = FALSE
)
}
\arguments{
\item{trial}{A CRT object or data frame containing (x,y) coordinates of
households}

\item{k}{integer: number of clusters in each arm}

\item{h}{integer: number of locations per cluster}

\item{algorithm}{algorithm for cluster boundaries, options are:
\code{NN} (the default),  \code{kmeans},  \code{TSP}}

\item{reuseTSP}{logical: indicator of whether a pre-existing path should be used by
the TSP algorithm}
}
\value{
A list of class \code{"CRTsp"} containing the following components:
\tabular{llll}{
\code{geom_full}   \tab list: \tab summary statistics describing the site,
and cluster assignments.\tab\cr
\code{trial} \tab data frame: \tab rows correspond to geolocated points, as follows:\tab\cr
\tab \code{x} \tab numeric vector: \tab x-coordinates of locations \cr
\tab \code{y} \tab numeric vector: \tab y-coordinates of locations \cr
\tab \code{cluster} \tab factor \tab assignments to cluster of each location  \cr
\tab \code{...} \tab \tab other objects included in the input \code{"CRTsp"} object or data frame \cr
}
}
\description{
\code{specify_clusters} algorithmically assigns locations to clusters by grouping them geographically
}
\details{
Clustering is carried out using one of three algorithms:
\tabular{lll}{
\code{NN}\tab Nearest neighbour:\tab assigns equal numbers of locations to each cluster \cr
\code{kmeans}\tab kmeans clustering  :\tab aims to partition locations so that each
belongs to the cluster with the nearest centroid \cr
\code{TSP} \tab travelling salesman problem heuristic :\tab Assigns locations sequentially
along a travelling salesman path.The \code{reuseTSP} parameter is used to allow the path to be reused
for creating alternative allocations with different cluster sizes. \cr
}
Either \code{k} or \code{h} must be specified. If both are specified
the input value of \code{k} is ignored.\cr
}
\examples{
#Assign clusters of average size h = 40 to a test set of co-ordinates, using the kmeans algorithm
exampletrial <- specify_clusters(trial = readdata('exampleCRT.txt'),
                            h = 40, algorithm = 'kmeans', reuseTSP = FALSE)
}
