% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Specify_CRTbuffer}
\alias{Specify_CRTbuffer}
\title{Specification of buffer zone in a cluster randomized trial}
\format{
data.frame:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
\item \code{cluster}: cluster assignment
\item \code{arm}: assignment to trial arm
\item \code{nearestDiscord}: distance to nearest discordant location
\item \code{buffer}: logical indicator of whether a point is in the buffer
}
}
\usage{
Specify_CRTbuffer(trial = trial, bufferWidth = 0)
}
\arguments{
\item{trial}{dataframe containing location coordinates and arm assignments}

\item{bufferWidth}{minimum distance between locations in the core areas of opposing arms}
}
\value{
The input dataframe augmented with distances to the nearest discordant location and a logical indicator of whether the location is in the buffer zone
}
\description{
\code{Specify_CRTbuffer} specifies a buffer zone in a cluster randomized trial (CRT) with geographical clustering by adding a flag to mark locations that are within a specified distance of locations in the opposite arm.
Locations are added incrementally to the buffer until a predetermined buffering distance is achieved.
}
\examples{
#Specify a buffer of 200m
exampletrial <- Specify_CRTbuffer(trial = readdata('test_Arms.csv'), bufferWidth = 0.2)

}
