% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.CRT}
\alias{summary.CRT}
\title{Summarise CRT}
\usage{
\method{summary}{CRT}(object, maskbuffer = 0.2, ...)
}
\arguments{
\item{object}{name of CRT}

\item{maskbuffer}{buffer in km around locations to use for calculating total area}

\item{...}{other arguments}
}
\description{
\code{summary.CRT} generates a summary description of a CRT object
}
