% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseCRT.R
\name{summary.CRTanalysis}
\alias{summary.CRTanalysis}
\title{Summarise analysis}
\usage{
\method{summary}{CRTanalysis}(object, ...)
}
\arguments{
\item{object}{name of analysis}

\item{...}{other arguments}
}
\description{
\code{summary.CRTanalysis} generates a summary description of an analysis of a CRT
}
