% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.CRTsp}
\alias{summary.CRTsp}
\title{Summary description of a \code{"CRTsp"} object}
\usage{
\method{summary}{CRTsp}(object, maskbuffer = 0.2, ...)
}
\arguments{
\item{object}{an object of class \code{"CRTsp"} or a data frame containing locations in (x,y) coordinates, cluster
assignments (factor \code{cluster}), and arm assignments (factor \code{arm}). Optionally: specification of a buffer zone (logical \code{buffer});
any other variables required for subsequent analysis.}

\item{maskbuffer}{radius of area around a location to include in calculation of areas}

\item{...}{other arguments used by summary}
}
\description{
\code{summary.CRTsp} provides a description of a \code{"CRTsp"} object
}
