% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_Analyse_CRT}
\alias{test_Analyse_CRT}
\title{Results of toy GEE analysis for use in testing}
\format{
list object with the following attributes:
\itemize{
\item \code{description}: Description of the trial dataset
\item \code{method}: statistical method
\item \code{PointEstimates}: point estimates
\item \code{IntervalEstimates}: interval estimates
}
}
\usage{
test_Analyse_CRT
}
\description{
Results of toy GEE analysis for use in testing
}
\keyword{datasets}
