% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_coordinates}
\alias{test_coordinates}
\title{Sample of coordinates from an arbitrary site in West Africa}
\format{
data.frame:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
})
}
\usage{
test_coordinates
}
\description{
Sample of coordinates from an arbitrary site in West Africa
}
\keyword{datasets}
