% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_design}
\alias{test_design}
\title{Test dataset containing the results of running DesignCRT() with the default options on the test_coordinates}
\format{
list object with the following attributes:
\itemize{
\item \code{arm}: vector of assignments to trial arms
\item \code{alpha}: confidence level
\item \code{power}: power
\item \code{seed}: random number seed
\item \code{effect}: Required effect size
\item \code{ICC}: Intra-Cluster Correlation obtained from other studies
\item \code{DE}: calculated Design Effect
\item \code{pC}: baseline prevalence
\item \code{cont}: contamination range in km, obtained from other studies
\item \code{coordinate_source}: filename for coordinates of households
\item \code{h}: proposal for the number of coordinates in each cluster
\item \code{algo}: algorithm used for cluster boundaries
\item \code{n_ind}: algorithm used for cluster boundaries
\item \code{assignments}: data frame containing locations and cluster assignments
\item \code{min_c}: minimum number of clusters required
}
}
\usage{
test_design
}
\description{
Test dataset containing the results of running DesignCRT() with the default options on the test_coordinates
}
\keyword{datasets}
