% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_result}
\alias{test_result}
\title{Dummy}
\format{
dataframe:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
}
}
\usage{
test_result
}
\description{
Dummy
}
\keyword{datasets}
