% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_Simulate_CRT}
\alias{test_Simulate_CRT}
\title{Test dataset containing the results of running Simulate_CRT() with the default options}
\format{
dataframe:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
\item \code{cluster}: cluster assignment
\item \code{arm}: assignment to trial arm
\item \code{infectiousness_proxy}: infectiousness proxy
\item \code{nearestDiscord}: distance to nearest discordant location
\item \code{p0}: value of underlying prevalence surface without intervention
\item \code{p1}: value of underlying prevalence surface with intervention
\item \code{denom}: number of samples evaluated at the location
\item \code{num}: number of positive samples
\item \code{index}: row number
}
}
\usage{
test_Simulate_CRT
}
\description{
Test dataset containing the results of running Simulate_CRT() with the default options
}
\keyword{datasets}
