% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_site}
\alias{test_site}
\title{Test dataset containing the results of running Simulate_TrialSite() with the default options}
\format{
data.frame:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
}
}
\usage{
test_site
}
\description{
Test dataset containing the results of running Simulate_TrialSite() with the default options
}
\keyword{datasets}
