% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testBuffer}
\alias{testBuffer}
\title{Trial dataframe containing buffer assignment based on Rusinga locations}
\format{
dataframe:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
\item \code{cluster}: cluster assignment
\item \code{arm}: assignment to trial arm
\item \code{nearestDiscord}: distance to nearest discordant location
\item \code{buffer}: logical indicator of whether a point is in the buffer
\item \code{index}: row number
}
}
\usage{
testBuffer
}
\description{
Trial dataframe containing buffer assignment based on Rusinga locations
}
\keyword{datasets}
