% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testClusters}
\alias{testClusters}
\title{Trial dataframe containing cluster assignments based on Rusinga locations}
\format{
dataframe:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
\item \code{cluster}: cluster assignment
}
}
\usage{
testClusters
}
\description{
Trial dataframe containing cluster assignments based on Rusinga locations
}
\keyword{datasets}
