% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testOutcomes}
\alias{testOutcomes}
\title{Trial data.frame including simulated outcomes based on Rusinga locations}
\format{
data.frame:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
\item \code{cluster}: cluster assignment
\item \code{arm}: assignment to trial arm
\item \code{infectiousness_proxy}: infectiousness proxy
\item \code{nearestDiscord}: distance to nearest discordant location
\item \code{p0}: value of underlying prevalence surface without intervention
\item \code{p1}: value of underlying prevalence surface with intervention
\item \code{buffer}: logical indicator of whether a point is in the buffer
\item \code{denom}: number of samples evaluated at the location
\item \code{num}: number of positive samples
\item \code{index}: row number
}
}
\usage{
testOutcomes
}
\description{
Trial data.frame including simulated outcomes based on Rusinga locations
}
\keyword{datasets}
