% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_CRT.R
\name{Analyse_CRT}
\alias{Analyse_CRT}
\title{Analysis of cluster randomized trial with contamination}
\usage{
Analyse_CRT(
  trial,
  method = "GEE",
  cfunc = "L",
  link = "logit",
  numerator = "num",
  denominator = "denom",
  excludeBuffer = FALSE,
  alpha = 0.05,
  baselineOnly = FALSE,
  baselineNumerator = "base_num",
  baselineDenominator = "base_denom",
  localisedEffects = FALSE,
  clusterEffects = FALSE,
  spatialEffects = FALSE,
  resamples = 1000,
  requireMesh = FALSE,
  inla.mesh = NULL
)
}
\arguments{
\item{trial}{trial dataframe including locations, clusters, arms, and outcomes}

\item{method}{statistical method used to analyse trial.
options are:
'EMP'  : empirical,
'T'   : comparison of cluster means by t-test
'GEE'  : Generalised Estimating Equations
'INLA' : INLA}

\item{cfunc}{transformation defining the contamination function
options are:
'S': piecewise linear (slope),
'L': inverse logistic (sigmoid),
'P': inverse probit,
'X': contamination not modelled'}

\item{link}{string: link function- options are 'logit' (the default), 'log', and 'identity'}

\item{numerator}{string: name of numerator variable for efficacy data (if present)}

\item{denominator}{string: name of denominator variable for efficacy data (if present)}

\item{excludeBuffer}{logical: indicator of whether any buffer zone (records with buffer=TRUE) should be excluded from analysis}

\item{alpha}{numeric: confidence level for confidence intervals and credible intervals}

\item{baselineOnly}{logical: indicator of whether required analysis is of efficacy or of baseline only}

\item{baselineNumerator}{string: name of numerator variable for baseline data (if present)}

\item{baselineDenominator}{string: name of denominator variable for baseline data (if present)}

\item{localisedEffects}{logical: indicator of whether the model includes local effects with no contamination}

\item{clusterEffects}{logical: indicator of whether the model includes cluster random effects}

\item{spatialEffects}{logical: indicator of whether the model includes spatial random effects}

\item{resamples}{integer: number of samples for sample-based intervals}

\item{requireMesh}{logical: indicator of whether spatial predictions are required}

\item{inla.mesh}{name of pre-existing INLA input object created by createMesh()}
}
\value{
list containing the following results of the analysis
\\itemize{
\\item \\code{description}: Description of the trial dataset
\\item \\code{method}: statistical method
\\item \\code{pt.ests}: point estimates
\\item \\code{int.ests}: interval estimates
\\item \\code{model.object}: object returned by the fitting function
\\item \\code{contamination}: function values and statistics describing the estimated contamination
}
}
\description{
\\code{Analyse_CRT} returns outputs from a statistical analysis of a cluster randomized trial (CRT).
}
\examples{
# Standard GEE analysis of test dataset ignoring contamination
exampleGEE=Analyse_CRT(trial=readdata('test_Simulate_CRT.csv'),method='GEE')
}
