% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_CRT.R
\name{Analyse_CRT}
\alias{Analyse_CRT}
\title{Analysis of cluster randomized trial with contamination}
\usage{
Analyse_CRT(
  trial,
  method = "sigmoid",
  excludeBuffer = FALSE,
  requireBootstrap = FALSE,
  alpha = 0.05,
  resamples = 1000,
  burnin = 500
)
}
\arguments{
\item{trial}{trial dataframe including locations, clusters, arms, and binary outcomes}

\item{method}{statistical method used to analyse trial. Options are 'piecewise_linear','logit','sigmoid','empirical','GEE','MCMC01','MCMC02','MCMC03'}

\item{excludeBuffer}{exclude any buffer zone (records with buffer=TRUE) from the analysis}

\item{requireBootstrap}{logical indicator of whether bootstrap confidence intervals are required}

\item{alpha}{confidence level for confidence intervals and credible intervals}

\item{resamples}{number of bootstrap samples}

\item{burnin}{number of burnin interations of MCMC algorithm (for MCMC methods)}
}
\value{
list containing the following results of the analysis
\itemize{
\item \code{description}: Description of the trial dataset
\item \code{method}: statistical method
\item \code{PointEstimates}: point estimates
\item \code{IntervalEstimates}: interval estimates
\item \code{ModelObject}: object returned by the fitting function
}
}
\description{
\code{Analyse_CRT} returns outputs from a statistical analysis of a cluster randomized trial (CRT).
}
\examples{
# Standard GEE analysis of test dataset
exampleGEE=Analyse_CRT(trial=test_Simulate_CRT,method='GEE')
}
