% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anonymize.site}
\alias{anonymize.site}
\title{Anonymize locations of a trial site}
\usage{
anonymize.site(trial)
}
\arguments{
\item{trial}{\code{"CRT"} object or trial data frame with Cartesian co-ordinates of
households (columns x and y)}
}
\value{
A list of class \code{"CRT"} containing the following components:
\tabular{lll}{
\code{CRT.design.full}   \tab list \tab summary statistics describing the site \cr
\code{x} \tab numeric vector \tab x-coordinates of locations \cr
\code{y} \tab numeric vector \tab y-coordinates of locations \cr
\code{...};   \tab other objects included in the input \code{"CRT"} object or data frame  \cr
}
}
\description{
\code{anonymize.site} carries out rotation of x,y coordinates by a random angle
about a random origin.
}
\details{
The returned object has transformed co-ordinates recentred at the origin. Other data are
unchanged.
}
\examples{
#Rotate and reflect test site locations
transformedTestlocations <- anonymize.site(trial =  readdata("test_Simulate_CRT.csv"))
}
