% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designCRT.R
\name{calculateCRTpower}
\alias{calculateCRTpower}
\title{Power and sample size calculations for a CRT}
\usage{
calculateCRTpower(
  locations,
  alpha,
  desiredPower,
  effect,
  yC,
  outcome.type,
  sigma2 = NULL,
  phi = 1,
  N = 1,
  ICC,
  mean_h,
  sd_h
)
}
\arguments{
\item{locations}{numeric: total number of units available for randomization}

\item{alpha}{numeric: confidence level}

\item{desiredPower}{numeric: desired power}

\item{effect}{numeric: required effect size}

\item{yC}{numeric: baseline outcome}

\item{outcome.type}{character with options-
\code{"y"}: continuous;
\code{"n"}: count;
\code{"e"}: event rate;
\code{"p"}: proportion;
\code{"d"}: dichotomous.}

\item{sigma2}{numeric: variance of the outcome (for \code{outcome.type = "y"})}

\item{phi}{numeric: overdispersion parameter (for \code{outcome.type = "n"} or \code{outcome.type = "e"})}

\item{N}{numeric: denominator for proportions (for \code{outcome.type = "p"})}

\item{ICC}{numeric: Intra-Cluster Correlation}

\item{mean_h}{mean number of units per cluster}

\item{sd_h}{standard deviation of number of units per cluster}
}
\value{
An object of class \code{"CRT"} comprising:
\itemize{
\item \code{design} list of input parameters
\item \code{geom.full} list containing the calculations of power, required number of clusters and design effect
}
}
\description{
\code{calculateCRTpower} carries out power and sample-size calculations for a CRT.
The output is an object of class \code{"CRT"} containing values for:
\itemize{
\item The required numbers of clusters to achieve the specified power.
\item The design effect based on the input ICC.
\item The power achievable with clusters of the specified size.
}
}
\details{
Power and sample size calculations are for a two-arm trial using the formulae of
\href{https://bmcmedresmethodol.biomedcentral.com/articles/10.1186/1471-2288-11-102}{Hemming et al, 2011}. For counts
or event rate data a quasi–Poisson model is assumed.
}
\examples{
examplePower = calculateCRTpower(locations = 3000, ICC=0.10, effect=0.4, alpha = 0.05,
    outcome.type = 'd', desiredPower = 0.8, yC=0.35, mean_h=100, sd_h=5)
}
