% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_CRT.R
\name{createMesh}
\alias{createMesh}
\title{\\code{createMesh} Create prediction mesh and other inputs required for INLA analyis of a CRT.}
\usage{
createMesh(
  trial = trial,
  offset = -0.1,
  max.edge = 0.25,
  inla.alpha = 2,
  maskbuffer = 0.5,
  ncells = 50
)
}
\arguments{
\item{trial}{trial dataframe including locations, clusters, arms, and binary outcomes}

\item{offset}{(see inla.mesh.2d documentation)}

\item{max.edge}{(see inla.mesh.2d documentation)}

\item{inla.alpha}{parameter related to the smoothness}

\item{maskbuffer}{(see inla.mesh.2d documentation)}

\item{ncells}{resolution of mesh in terms of maximum of linear dimension in pixels}
}
\value{
list containing the mesh
\\itemize{
\\item \\code{prediction}: Data.table containing the prediction points and covariate values
\\item \\code{A}: projection matrix from the observations to the mesh nodes.
\\item \\code{Ap}: projection matrix from the prediction points to the mesh nodes.
\\item \\code{indexs}:  index set for the SPDE model
\\item \\code{spde}: SPDE model
}
}
\description{
\\code{createMesh} Create prediction mesh and other inputs required for INLA analyis of a CRT.
}
\examples{
# low resolution mesh for test dataset
exampleMesh=createMesh(trial=test_Simulate_CRT,ncells=10)
}
