% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTspat-package.R
\docType{package}
\name{CRTspat-package}
\alias{CRTspat}
\alias{CRTspat-package}
\title{CRTspat: Design and Analysis of Cluster Randomized Trials of Health Interventions with Spatial Contamination}
\description{
functions to support design, workflow and analysis for Cluster Randomised Trials of (health) interventions where there may be geographical contamination between the arms. cluster size and sample size for cluster randomized controls of interventions with geographical contamination, and for carrying out randomizations for such trials. The methods are described in Multerer et al. (in preparation). The package takes as inputs a table of geolocations of experimental units. Example datasets are provided taken from field trials of malaria interventions.
}
\author{
\strong{Maintainer}: Thomas Smith \email{thomas-a.smith@unibas.ch} (\href{https://orcid.org/0000-0002-3650-9381}{ORCID}) [copyright holder]

}
\keyword{internal}
