% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design_CRT.R
\name{Design_CRT}
\alias{Design_CRT}
\title{Design a CRT of a malaria intervention with contamination}
\usage{
Design_CRT(
  alpha = 0.05,
  power = 0.8,
  effect = 0.4,
  ICC = 0.175,
  pC = 0.4,
  cont = 0.25,
  coordinates = AvecNet_coordinates,
  h = 80,
  algo = "kmeans",
  reuseTSP = FALSE
)
}
\arguments{
\item{alpha}{confidence level}

\item{power}{power}

\item{effect}{required effect size}

\item{ICC}{Intra-Cluster Correlation obtained from other studies}

\item{pC}{baseline prevalence}

\item{cont}{contamination range in km, obtained from other studies}

\item{coordinates}{dataframe containing coordinates of households. Columns 'x' and 'y' should contain Cartesian (x,y) coordinates. Units are expected to be km.}

\item{h}{proposal for the number of coordinates in each cluster}

\item{algo}{algorithm for cluster boundaries, choose between
"TSP": travelling salesman problem heuristic;
"NN": nearest neighbor;
"kmeans": kmeans algorithm}

\item{reuseTSP}{indicator of whether a pre-existing path should be used by the TSP algorithm}
}
\value{
A trial design object comprising a list with the following attributes:
\itemize{
\item \code{arm}: vector of assignments to trial arms
\item \code{alpha}: confidence level
\item \code{power}: power
\item \code{effect}: Required effect size
\item \code{ICC}: Intra-Cluster Correlation obtained from other studies
\item \code{DE}: calculated Design Effect
\item \code{pC}: baseline prevalence
\item \code{cont}: contamination range in km, obtained from other studies
\item \code{coordinates}: dataframe containing (x,y) coordinates of locations
\item \code{h}: proposal for the number of households in each cluster
\item \code{algo}: algorithm used for cluster boundaries
\item \code{assignments}: data frame containing locations, clusters and arm assignments
\item \code{min_c}: minimum number of clusters required
}
}
\description{
\code{Design_CRT} estimates the required number of clusters and the extent of contamination between arms for a CRT with the input set of locations.
Outputs are:
(i) Estimates of the required numbers of clusters.
(ii) A proposal for the cluster and arm assignments to the input coordinates. (A warning is output if the number of locations is too small to allow randomisation of sufficient clusters).
(iii) the proportion of households in the input geography falling within the core of the clusters (i.e. outside the contamination range of locations in the opposite arm)
}
\examples{

exampleDesign = Design_CRT(coordinates=test_site, ICC=0.10, effect=0.4, pC=0.35, cont=0.5, h=100)
}
