% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designCRT.R
\name{designCRT}
\alias{designCRT}
\title{Design a CRT for a given set of locations}
\usage{
designCRT(
  trial,
  alpha = 0.05,
  desiredPower = 0.8,
  effect,
  yC,
  outcome.type = "d",
  sigma2 = NULL,
  phi = 1,
  N = 1,
  ICC,
  buffer.width = 0,
  h,
  algorithm = "kmeans"
)
}
\arguments{
\item{trial}{dataframe or \code{"CRT"} object containing Cartesian coordinates of locations in columns 'x' and 'y'.
Units are expected to be km.}

\item{alpha}{numeric: confidence level}

\item{desiredPower}{numeric: desired power}

\item{effect}{numeric: required effect size}

\item{yC}{numeric: baseline outcome}

\item{outcome.type}{character with options -
\code{"y"}: continuous;
\code{"n"}: count;
\code{"e"}: event rate;
\code{"p"}: proportion;
\code{"d"}: dichotomous.}

\item{sigma2}{numeric: variance of the outcome (for \code{outcome.type = "y"})}

\item{phi}{numeric: overdispersion parameter (for \code{outcome.type = "n"} or \code{outcome.type = "e"})}

\item{N}{numeric: denominator for proportions (for \code{outcome.type = "p"})}

\item{ICC}{numeric: Intra-Cluster Correlation}

\item{buffer.width}{numeric: required buffer width in km}

\item{h}{numeric: proposal for the number of coordinates in each cluster}

\item{algorithm}{algorithm used to determine cluster boundaries, options are
\code{"TSP"}: travelling salesman problem heuristic;
\code{"NN"}: nearest neighbor;
\code{"kmeans"}: kmeans.}
}
\value{
A \code{"CRT"} object comprising the input data, cluster and arm assignments, trial description and results of power calculations
\tabular{lll}{
\code{geom.full}   \tab list \tab summary statistics describing the site,
cluster assignments, and randomization.\cr
\code{x} \tab numeric vector \tab x-coordinates of locations \cr
\code{y} \tab numeric vector \tab y-coordinates of locations \cr
\code{cluster} \tab factor \tab assignments to cluster of each location  \cr
\code{arm} \tab factor \tab assignments to \code{"control"} or \code{"intervention"} for each location \cr
\code{buffer} \tab logical \tab indicator of whether the point is within the buffer \cr
\code{...} \tab other objects included in the input \code{"CRT"} object or data frame \cr
}
}
\description{
\code{designCRT} estimates the required number of clusters and the extent of contamination between arms for a CRT
with the input set of locations. Outputs are:
(i) Estimates of the required numbers of clusters.
(ii) A proposal for the cluster and arm assignments to the input coordinates.
(A warning is output if the number of locations is too small to allow randomisation of sufficient clusters).
(iii) the proportion of households in the input geography falling within the core of the clusters.
}
\examples{
exampleDesign = designCRT(trial = readdata('test_site.csv'),
                desiredPower = 0.8, effect=0.4, yC=0.35, outcome.type = 'd', ICC = 0.05,
                buffer.width = 0.25, h=100, algorithm = 'kmeans')
}
