% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readdata}
\alias{readdata}
\title{Read example dataset}
\usage{
readdata(filename)
}
\arguments{
\item{filename}{name of text file stored within the package}
}
\value{
R object corresponding to the text file
}
\description{
\code{readdata} reads a file from the package library of example datasets
}
\details{
The input file name should include the extension (either .csv or .txt).
The resulting object is a data frame if the extension is .csv.
}
\examples{
exampleCRT <- readdata('exampleCRT.txt')

}
