% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_CRT.R
\name{Simulate_CRT}
\alias{Simulate_CRT}
\title{Simulation of cluster randomized trial with contamination}
\usage{
Simulate_CRT(
  trial = NULL,
  efficacy = 0,
  initialPrevalence = NULL,
  generateBaseline = TRUE,
  matchedPair = TRUE,
  baselineNumerator = "base_num",
  baselineDenominator = "base_denom",
  denominator = NULL,
  ICC_inp = NULL,
  sd = NULL,
  theta_inp = NULL,
  tol = 1e-04
)
}
\arguments{
\item{trial}{a dataframe containing locations (x,y), cluster assignments, and arm assignments}

\item{efficacy}{simulated efficacy (defaults to 0)}

\item{initialPrevalence}{prevalence in control arm (assumed equal to initial proportion)}

\item{generateBaseline}{logical indicator of whether baseline data should be simulated}

\item{matchedPair}{logical: indicator of whether pair-matching on the baseline data should be used in randomization}

\item{baselineNumerator}{name of numerator variable for baseline data (if present)}

\item{baselineDenominator}{name of denominator variable for baseline data (if present)}

\item{denominator}{name of denominator variable for the outcome (if present)}

\item{ICC_inp}{Intra Cluster Correlation, provided as input when baseline data are to be simulated}

\item{sd}{standard deviation of the normal kernel measuring spatial smoothing leading to contamination}

\item{theta_inp}{input contamination range}

\item{tol}{tolerance of output ICC}
}
\value{
A trial simulation object comprising a data frame containing the following numerical quantities:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
\item \code{cluster}: assignment to cluster
\item \code{arm}: assignment to trial arm
\item \code{propensity}: propensity of location
\item \code{base_denom}: denominator for baseline
\item \code{base_num}: numerator for baseline
\item \code{nearestDiscord}: distance to nearest discordant location
\item \code{denom}: denominator for the outcome
\item \code{num}: numerator for the outcome
}
}
\description{
\code{Simulate_CRT} generates simulated data for a cluster randomized trial (CRT) with geographic contamination between arms and a binary outcome. Contamination is simulated as arising from a diffusion-like process.
For details see Multerer (PhD thesis).
}
\examples{
example_simulated_CRT =  Simulate_CRT(trial=readdata('test_Arms.csv'),
                                     efficacy=0.25,
                                     ICC_inp=0.05,
                                     initialPrevalence=0.5,
                                     matchedPair = TRUE,
                                     sd=0.6,
                                     tol=0.05)
}
