% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{specify.buffer}
\alias{specify.buffer}
\title{Specification of buffer zone in a cluster randomized trial}
\usage{
specify.buffer(trial, buffer.width = 0)
}
\arguments{
\item{trial}{an object of class \code{"CRT"} or a data frame containing locations in (x,y) coordinates, cluster
assignments (factor \code{cluster}), and arm assignments (factor \code{arm}).}

\item{buffer.width}{minimum distance between locations in
opposing arms for them to qualify to be included in the core area (km)}
}
\value{
an object of class \code{"CRT"} comprising the following components:
\tabular{lll}{
\code{CRT.design.full}   \tab list \tab summary statistics describing the site,
cluster assignments, and randomization.\cr
\code{CRT.design.core}   \tab list \tab summary statistics describing the core area \cr
\code{x} \tab numeric vector \tab x-coordinates of locations \cr
\code{y} \tab numeric vector \tab y-coordinates of locations \cr
\code{cluster} \tab factor \tab assignments to cluster of each location  \cr
\code{arm} \tab factor \tab assignments to \code{"control"} or \code{"intervention"} for each location \cr
\code{nearestDiscord} \tab numeric vector \tab Euclidean distance to nearest discordant location (km) \cr
\code{buffer} \tab logical \tab indicator of whether the point is within the buffer \cr
\code{...} \tab other objects included in the input \code{"CRT"} object or data frame \cr
}
}
\description{
\code{specify.buffer} specifies a buffer zone in a cluster randomized
trial (CRT) by flagging those locations that are within a defined distance of
those in the opposite arm.
}
\examples{
#Specify a buffer of 200m
exampletrial <- specify.buffer(trial = readdata('test_Arms.csv'), buffer.width = 0.2)
}
