% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseCRT.R
\name{summary.CRTanalysis}
\alias{summary.CRTanalysis}
\title{Summary of the results of a statistical analysis of a CRT}
\usage{
\method{summary}{CRTanalysis}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"CRTanalysis"}}

\item{...}{other arguments used by summary}
}
\value{
No return value, writes text to the console.
}
\description{
\code{summary.CRTanalysis} generates a summary of a \code{CRTanalysis} including the main results
}
\examples{
{example <- readdata('exampleCRT.txt')
exampleT <- CRTanalysis(example, method = "T")
summary(exampleT)
}
}
