% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTsp.R
\name{summary.CRTsp}
\alias{summary.CRTsp}
\title{Summary description of a \code{"CRTsp"} object}
\usage{
\method{summary}{CRTsp}(object, maskbuffer = 0.2, ...)
}
\arguments{
\item{object}{an object of class \code{"CRTsp"} or a data frame containing locations in (x,y) coordinates, cluster
assignments (factor \code{cluster}), arm assignments (factor \code{arm}) and buffer zones (logical \code{buffer}),
together with any other variables required for subsequent analysis.}

\item{maskbuffer}{radius of area around a location to include in calculation of areas}

\item{...}{other arguments used by summary}
}
\value{
No return value, write text to the console.
}
\description{
\code{summary.CRTsp} provides a description of a \code{"CRTsp"} object
}
\examples{
summary(CRTsp(readdata('exampleCRT.txt')))
}
