% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testArms}
\alias{testArms}
\title{Trial data.frame containing arm assignments based on Rusinga locations}
\format{
data.frame:
\itemize{
\item \code{x}: x-coordinates of location
\item \code{y}: y-coordinates of location
\item \code{cluster}: cluster assignment
\item \code{arm}: assignment to trial arm
}
}
\usage{
testArms
}
\description{
Trial data.frame containing arm assignments based on Rusinga locations
}
\keyword{datasets}
